New meny entry, but no real action for now
Aim: shortcut in order to allow a preselected browser to bypass Tor and
the firewall.
If you click on it, it will punch a hole in iptables, and if possible
launch the browser on some neutral address in order to be caught by the
captive portal.
We will need a way in order to close the hole asap. Maybe once we quit
the browser instance… Or a timer…