Allow a special, selected browser to bypass Tor and Firewall
In order to log in some captive portal, such a hack is needed. This part
opens the firewall for 5 minutes for a user selected application, then
close it again.
You may, of course, disable the rule directly from the menu once you're
identified on the captive portal.
You will get a Toast every 30 seconds as a reminder. It displays
remaining time so that you know what to expect.