Reset browser authorization status
@patcon reported it in #17.

Technical aspect:
The authorization state wasn't reset properly if the device was rebooted
during the grace time.

The timer is stopped when we want to disable the browser before
gracetime end. Browser state wasn't correct after the reboot, meaning we
wanted to "deactivate browser" when it wasn't activated, thus stopping
a timer that was never started.

This didn't affect orwall security (iptables rules were applied as
expected).