Set "select all" as "never shown as Action"
This option stinks, and will probably be removed. Or maybe it will
require some activation in the settings — this "select all" shouldn't be
used. At all. If more than half of your apps requires an access, it may
be good to think a bit about their aim, and you use-case.