Removed "select all" menu entry and related stuff
This won't be of any use with the coming features, and the implem was
really crappy.
Also, I wasn't convince by the use case. If people really does have so
many apps accessing the the Net, using tor won't be really useful: each
app accessing to the Net may give away your ID.