New classes in order to use SQLite as backend for NAT rules
The current StringSet in SharedPreferences cannot comply with the new
features we'll add to orWall.
Using SQLite will allow us to provide new features like per-app Onion
Router choice, ports and so on.

Most probably the schema will be updated shortly, it just shows the
basis. Most of the fields will not be used for now, as we need some more
time in order to modify the UI and provide new options.

SharedPreferences is kept in order to set application preferences only
(stuff in settings).