Modified NetworkHelper in order to avoid useless permission
The "android.permission.INTERNET" permission is pretty useless for
orWall, as we don't need Internet access. At all. We just need
connection information.

The NetworkInterface API call seems to require the INTERNET permission.
Using the WifiManager allows us to remove this requirement and just
stick with a smaller permissions set:
- ACCESS_WIFI_STATE
- ACCESS_NETWORK_STATE