Allow LAN bypass to be triggered from NetworkReceiver
In previous patch, a bug was fixed. This new patch allow orWall to
manage in a better way the LAN bypass (though it's still not perfect).

Next step would be: in case we don't get any IP on the Wifi, we should
ensure no deprecated rule stay in IPTables. This means we should be able
to register the LAN subnet in order to take it back.
This could be done un the sharedPreferences.