BootBroadcast must ensure we have the init-script installed
Doing so will allow us to ensure, on each boot, the init-script is
present.
This will NOT correct issue #46. This one is a bit nastier, as it seems
some ROMs have issue with the userinit.d location. We suspect crypted
devices to cause some problems at this level (maybe the data directory
isn't present early enough).