Prevent orWall to call root on each network change
This should prevent many popups regarding orWall root accesses.
The small problem now: we get the subnet in the sharedPreferences. In
plain.
This MAY lead to the leak of some network information if some other
application get a hand on the sharePreference content, but I think this
won't be the main problem if such a case happens…