New type: PackageInfoData
This is a small meta-type, creating a simple objet holding some package
information.
This is used especially for special apps like Media Server, as they
don't show up "naturally" in orWall.
Code is strongly based (well, copied, sorry) on AFWall+ API library.
Thanks a lot to them for the work — hopefully we will be able to
collaborate on their new version, in order to provide some intents on
their side so that we can remove a lot of things from orWall in a not so
far future.