Corrected stupid bug
When we unchecked an app, it didn't take care of its real configuration,
and just removed standard, default "redirect" version.
Thus, when we checked something special in Advanced Settings, the rules
weren't properly removed unless we reboot the device.

Now, we take care of that, and we also take the opportunity to replace
dynamically objects in "apps" listing with latest configurations.